+++ server/s3_audit_info_logger_iem.h
@@ -32,8 +32,8 @@ class S3AuditInfoLoggerIEM {
   typedef struct event_base evbase_t;
 
  public:
-  S3AuditInfoLoggerIEM(evbase_t* p_base, std::string host_ip,
-                            long port, std::string path);
+  S3AuditInfoLoggerIEM(evbase_t* p_base, std::string host_ip, uint16_t port,
+                       std::string path);
   int save_msg(std::string const& eventID,
               std::string const& severity,
               std::string const& msg);
